Rename core_transforms/{CoGroupByKey/CoGroupByKey -> cogroupbykey/cogroupbykey}
