+++ learning/katas/go/core_transforms/cogroupbykey/cogroupbykey/cmd/main.go
@@ -13,17 +13,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-module cogroupbykey
+package main
 
-go 1.14
-
-require (
-	github.com/apache/beam v2.22.0+incompatible
-	github.com/golang/protobuf v1.4.2 // indirect
-	github.com/google/go-cmp v0.4.1 // indirect
-	golang.org/x/net v0.0.0-20200602114024-627f9648deb9 // indirect
-	golang.org/x/sys v0.0.0-20200610111108-226ff32320da // indirect
-	google.golang.org/api v0.25.0 // indirect
-	google.golang.org/genproto v0.0.0-20200611194920-44ba362f84c1 // indirect
-	google.golang.org/grpc v1.29.1 // indirect
+import (
+	"beam.apache.org/learning/katas/core_transforms/cogroupbykey/cogroupbykey/pkg/task"
+	"context"
+	"github.com/apache/beam/sdks/go/pkg/beam"
+	"github.com/apache/beam/sdks/go/pkg/beam/log"
+	"github.com/apache/beam/sdks/go/pkg/beam/x/beamx"
+	"github.com/apache/beam/sdks/go/pkg/beam/x/debug"
 )
+
+func main() {
+	ctx := context.Background()
+
+	p, s := beam.NewPipelineWithRoot()
+
+	fruits := beam.Create(s.Scope("Fruits"), "apple", "banana", "cherry")
+	countries := beam.Create(s.Scope("Countries"), "australia", "brazil", "canada")
+
+	output := task.ApplyTransform(s, fruits, countries)
+
+	debug.Print(s, output)
+
+	err := beamx.Run(ctx, p)
+
+	if err != nil {
+		log.Exitf(ctx, "Failed to execute job: %v", err)
+	}
+}
