+++ learning/katas/go/core_transforms/cogroupbykey/cogroupbykey/task-remote-info.yaml
