+++ learning/katas/go/core_transforms/cogroupbykey/cogroupbykey/task.md
