+++ learning/katas/go/core_transforms/cogroupbykey/lesson-remote-info.yaml
