+++ learning/katas/go/core_transforms/section-info.yaml
@@ -20,7 +20,7 @@
 content:
 - map
 - groupbykey
-- CoGroupByKey
+- cogroupbykey
 - Combine
 - Flatten
 - Partition
